+++ apps/backend/src/modules/notifications/notification.repository.ts
@@ -1,4 +1,3 @@
-import { EMPTY_LENGTH } from "~/libs/constants/constants.js";
 import { RelationName, SortOrder } from "~/libs/enums/enums.js";
 import { type Repository, type ValueOf } from "~/libs/types/types.js";
 import { type UserModel } from "~/modules/users/users.js";
@@ -31,25 +30,14 @@ class NotificationRepository implements Repository<NotificationEntity> {
 		return notificationTypeToMessage[type];
 	}
 
-	public async checkHasUserUnreadNotifications(
-		userId: number,
-	): Promise<boolean> {
-		const unreadNotifications = await this.notificationModel
-			.query()
-			.where({ receiverUserId: userId })
-			.andWhere({ status: NotificationStatus.UNREAD });
-
-		return unreadNotifications.length > EMPTY_LENGTH;
-	}
-
 	public async create(
 		payload: NotificationEntity,
 	): Promise<NotificationEntity> {
-		const { receiverUserId, type, userId } = payload.toNewObject();
+		const { actionId, receiverUserId, type, userId } = payload.toNewObject();
 
 		const createdNotification = await this.notificationModel
 			.query()
-			.insert({ receiverUserId, type, userId })
+			.insert({ actionId, receiverUserId, type, userId })
 			.returning("*")
 			.withGraphFetched(
 				`${RelationName.USER}.${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
@@ -57,6 +45,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 			.execute();
 
 		return NotificationEntity.initialize({
+			actionId: createdNotification.actionId,
 			createdAt: createdNotification.createdAt,
 			id: createdNotification.id,
 			message: this.getMessageByType(
@@ -84,6 +73,19 @@ class NotificationRepository implements Repository<NotificationEntity> {
 		return Boolean(deletedRowsCount);
 	}
 
+	public async deleteByActionId(
+		actionId: number,
+		type: ValueOf<typeof NotificationType>,
+	): Promise<boolean> {
+		const deletedItemsCount = await this.notificationModel
+			.query()
+			.where({ actionId, type })
+			.delete()
+			.execute();
+
+		return Boolean(deletedItemsCount);
+	}
+
 	public async find(
 		notificationId: number,
 	): Promise<NotificationEntity | null> {
@@ -97,6 +99,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 
 		return notification
 			? NotificationEntity.initialize({
+					actionId: notification.actionId,
 					createdAt: notification.createdAt,
 					id: notification.id,
 					message: this.getMessageByType(notification.type, notification.user),
@@ -122,6 +125,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 
 		return notifications.map((notification) => {
 			return NotificationEntity.initialize({
+				actionId: notification.actionId,
 				createdAt: notification.createdAt,
 				id: notification.id,
 				message: this.getMessageByType(notification.type, notification.user),
@@ -169,6 +173,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 
 		return userNotifications.map((notification) => {
 			return NotificationEntity.initialize({
+				actionId: notification.actionId,
 				createdAt: notification.createdAt,
 				id: notification.id,
 				message: this.getMessageByType(notification.type, notification.user),
@@ -182,6 +187,48 @@ class NotificationRepository implements Repository<NotificationEntity> {
 				userLastName: notification.user.userDetails.lastName,
 			});
 		});
+	}
+
+	public async findByActionId(
+		actionId: number,
+		type: ValueOf<typeof NotificationType>,
+	): Promise<NotificationEntity | null> {
+		const notification = await this.notificationModel
+			.query()
+			.where({ actionId, type })
+			.withGraphFetched(
+				`${RelationName.USER}.${RelationName.USER_DETAILS}.${RelationName.AVATAR_FILE}`,
+			)
+			.first();
+
+		return notification
+			? NotificationEntity.initialize({
+					actionId: notification.actionId,
+					createdAt: notification.createdAt,
+					id: notification.id,
+					message: this.getMessageByType(notification.type, notification.user),
+					receiverUserId: notification.receiverUserId,
+					status: notification.status,
+					type: notification.type,
+					updatedAt: notification.updatedAt,
+					userAvatarUrl: notification.user.userDetails.avatarFile?.url ?? null,
+					userFirstName: notification.user.userDetails.firstName,
+					userId: notification.userId,
+					userLastName: notification.user.userDetails.lastName,
+				})
+			: null;
+	}
+
+	public async getUnreadNotificationsCount(userId: number): Promise<number> {
+		const { count } = await this.notificationModel
+			.query()
+			.where({ receiverUserId: userId })
+			.andWhere({ status: NotificationStatus.UNREAD })
+			.count()
+			.first()
+			.castTo<{ count: number }>();
+
+		return count;
 	}
 
 	public async setReadNotifications(
@@ -199,6 +246,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 
 		return updatedNotifications.map((notification) => {
 			return NotificationEntity.initialize({
+				actionId: notification.actionId,
 				createdAt: notification.createdAt,
 				id: notification.id,
 				message: this.getMessageByType(notification.type, notification.user),
@@ -228,6 +276,7 @@ class NotificationRepository implements Repository<NotificationEntity> {
 			.execute();
 
 		return NotificationEntity.initialize({
+			actionId: updatedNotification.actionId,
 			createdAt: updatedNotification.createdAt,
 			id: updatedNotification.id,
 			message: this.getMessageByType(