+++ apps/backend/src/modules/user-notifications/user-notification.controller.ts
@@ -83,14 +83,14 @@ class UserNotificationController extends BaseController {
 
 		this.addRoute({
 			handler: (options) => {
-				return this.checkHasUserUnreadNotifications(
+				return this.getUnreadNotificationsCount(
 					options as APIHandlerOptions<{
 						user: UserAuthResponseDto;
 					}>,
 				);
 			},
 			method: "GET",
-			path: UserNotificationsApiPath.CHECK_HAS_USER_UNREAD_NOTIFICATIONS,
+			path: UserNotificationsApiPath.UNREAD_COUNT,
 		});
 
 		this.addRoute({
@@ -105,36 +105,6 @@ class UserNotificationController extends BaseController {
 				body: readNotificationsRequestValidationSchema,
 			},
 		});
-	}
-
-	/**
-	 * @swagger
-	 * /user-notifications/check-has-user-unread-notifications:
-	 *    get:
-	 *      tags:
-	 *        - User notifications
-	 *      security:
-	 *        - bearerAuth: []
-	 *      description: Checks if user has unread notifications
-	 *      responses:
-	 *        200:
-	 *          description: Successful operation
-	 *          content:
-	 *            application/json:
-	 *              schema:
-	 *                type: boolean
-	 */
-	public async checkHasUserUnreadNotifications(
-		options: APIHandlerOptions<{
-			user: UserAuthResponseDto;
-		}>,
-	): Promise<APIHandlerResponse> {
-		return {
-			payload: await this.notificationService.checkHasUserUnreadNotifications(
-				options.user.id,
-			),
-			status: HTTPCode.OK,
-		};
 	}
 
 	/**
@@ -181,6 +151,36 @@ class UserNotificationController extends BaseController {
 				options.user.id,
 				options.query.type ?? NotificationFilter.ALL,
 				options.query.search ?? "",
+			),
+			status: HTTPCode.OK,
+		};
+	}
+
+	/**
+	 * @swagger
+	 * /user-notifications/unread-count:
+	 *    get:
+	 *      tags:
+	 *        - User notifications
+	 *      security:
+	 *        - bearerAuth: []
+	 *      description: Get a count of all user's unread notifications
+	 *      responses:
+	 *        200:
+	 *          description: Successful operation
+	 *          content:
+	 *            application/json:
+	 *              schema:
+	 *                type: number
+	 */
+	public async getUnreadNotificationsCount(
+		options: APIHandlerOptions<{
+			user: UserAuthResponseDto;
+		}>,
+	): Promise<APIHandlerResponse> {
+		return {
+			payload: await this.notificationService.getUnreadNotificationsCount(
+				options.user.id,
 			),
 			status: HTTPCode.OK,
 		};