+++ apps/frontend/src/libs/components/header/libs/constants/constants.ts
@@ -1,0 +1,1 @@
+export { MAXIMUM_DISPLAY_UNREAD_COUNT } from "./maximum-display-unread-count.constant.js";